move project geom styles to util to avoid duplication
